+++ apps/backend/src/modules/points-of-interest/points-of-interest.entity.ts
@@ -1,5 +1,7 @@
 import { type Entity, type PointGeometry } from "~/libs/types/types.js";
 
+import { RouteEntity } from "../routes/route.entity.js";
+
 class PointsOfInterestEntity implements Entity {
 	private createdAt: null | string;
 
@@ -11,24 +13,29 @@ class PointsOfInterestEntity implements Entity {
 
 	private name: string;
 
+	private routes: null | ReturnType<RouteEntity["toObject"]>[];
+
 	private constructor({
 		createdAt,
 		description,
 		id,
 		location,
 		name,
+		routes = null,
 	}: {
 		createdAt: null | string;
 		description: null | string;
 		id: null | number;
 		location: PointGeometry;
 		name: string;
+		routes: null | ReturnType<RouteEntity["toObject"]>[];
 	}) {
 		this.id = id;
 		this.location = location;
 		this.name = name;
 		this.createdAt = createdAt;
 		this.description = description;
+		this.routes = routes;
 	}
 
 	public static initialize(data: {
@@ -37,6 +44,8 @@ class PointsOfInterestEntity implements Entity {
 		id: number;
 		location: PointGeometry;
 		name: string;
+		routes: null | ReturnType<RouteEntity["toObject"]>[];
+		updatedAt: string;
 	}): PointsOfInterestEntity {
 		return new PointsOfInterestEntity({
 			createdAt: data.createdAt,
@@ -44,6 +53,9 @@ class PointsOfInterestEntity implements Entity {
 			id: data.id,
 			location: data.location,
 			name: data.name,
+			routes:
+				data.routes?.map((route) => RouteEntity.initialize(route).toObject()) ??
+				[],
 		});
 	}
 
@@ -62,7 +74,26 @@ class PointsOfInterestEntity implements Entity {
 			id: null,
 			location,
 			name,
+			routes: null,
 		});
+	}
+
+	public toDetailsObject(): {
+		description: null | string;
+		id: number;
+		location: PointGeometry;
+		name: string;
+		routes: { id: number; name: string }[];
+	} {
+		return {
+			description: this.description,
+			id: this.id as number,
+			location: this.location,
+			name: this.name,
+			routes: this.routes
+				? this.routes.map(({ id, name }) => ({ id, name }))
+				: [],
+		};
 	}
 
 	public toListObject(): {