+++ apps/backend/src/modules/user-routes/libs/types/type.ts
@@ -1,0 +1,7 @@
+export {
+	type UserRouteCreateRequestDto,
+	type UserRouteParameters,
+	type UserRoutePatchRequestDto,
+	type UserRouteResponseDto,
+	type UserRouteStatusType,
+} from "@smartscapes/shared";